Move demo files to utilize nblink with sphinx
