+++ demos/data/digitized_plant.geojson
@@ -1,0 +1,17 @@
+{
+"type": "FeatureCollection",
+"name": "sample_plant",
+"crs": { "type": "name", "properties": { "name": "urn:ogc:def:crs:OGC:1.3:CRS84" } },
+"features": [
+{ "type": "Feature", "properties": { "id": null, "asset_type": "combiner", "asset_name": "CMB-001-01" }, "geometry": { "type": "MultiPolygon", "coordinates": [ [ [ [ -83.673025205823308, 33.680229416591168 ], [ -83.673025205823308, 33.680069896249719 ], [ -83.673667392351959, 33.68007189025581 ], [ -83.673664996133596, 33.680109776362826 ], [ -83.673449336478427, 33.680113764373111 ], [ -83.673449336478427, 33.680231410593564 ], [ -83.673025205823308, 33.680229416591168 ] ] ] ] } },
+{ "type": "Feature", "properties": { "id": null, "asset_type": "combiner", "asset_name": "CMB-001-02" }, "geometry": { "type": "MultiPolygon", "coordinates": [ [ [ [ -83.672821527260098, 33.680055938205761 ], [ -83.672821527260098, 33.679776776850709 ], [ -83.674093919225484, 33.679780764876455 ], [ -83.674091523007064, 33.679938291745209 ], [ -83.673875863351924, 33.679940285754356 ], [ -83.673875863351924, 33.680055938205761 ], [ -83.672821527260098, 33.680055938205761 ] ] ] ] } },
+{ "type": "Feature", "properties": { "id": null, "asset_type": "combiner", "asset_name": "CMB-001-03" }, "geometry": { "type": "MultiPolygon", "coordinates": [ [ [ [ -83.672608263823363, 33.679752848692395 ], [ -83.672608263823363, 33.679477674393034 ], [ -83.674089126788701, 33.679477674393034 ], [ -83.674089126788701, 33.679752848692395 ], [ -83.672608263823363, 33.679752848692395 ] ] ] ] } },
+{ "type": "Feature", "properties": { "id": null, "asset_type": "combiner", "asset_name": "CMB-001-04" }, "geometry": { "type": "MultiPolygon", "coordinates": [ [ [ [ -83.672404585260153, 33.679457734192198 ], [ -83.672404585260153, 33.679178570894834 ], [ -83.674086730570309, 33.679178570894834 ], [ -83.674086730570309, 33.67946172223273 ], [ -83.672404585260153, 33.679457734192198 ] ] ] ] } },
+{ "type": "Feature", "properties": { "id": null, "asset_type": "combiner", "asset_name": "CMB-001-05" }, "geometry": { "type": "MultiPolygon", "coordinates": [ [ [ [ -83.672196114260217, 33.679160624651843 ], [ -83.672196114260217, 33.679001102327376 ], [ -83.674084334351946, 33.679001102327376 ], [ -83.674084334351946, 33.67916261867903 ], [ -83.672196114260217, 33.679160624651843 ] ] ] ] } },
+{ "type": "Feature", "properties": { "id": null, "asset_type": "combiner", "asset_name": "CMB-001-06" }, "geometry": { "type": "MultiPolygon", "coordinates": [ [ [ [ -83.672199708587812, 33.678976176937439 ], [ -83.672199708587812, 33.678825627428608 ], [ -83.674077145696771, 33.678825627428608 ], [ -83.674077145696771, 33.678983156047359 ], [ -83.672199708587812, 33.678976176937439 ] ] ] ] } },
+{ "type": "Feature", "properties": { "id": null, "asset_type": "combiner", "asset_name": "CMB-001-07" }, "geometry": { "type": "MultiPolygon", "coordinates": [ [ [ [ -83.671980454605091, 33.678799704970025 ], [ -83.671980454605091, 33.678592325019991 ], [ -83.674089126788715, 33.678592325019991 ], [ -83.674089126788715, 33.678803693041097 ], [ -83.671980454605091, 33.678799704970025 ] ] ] ] } },
+{ "type": "Feature", "properties": { "id": null, "asset_type": "combiner", "asset_name": "CMB-001-08" }, "geometry": { "type": "MultiPolygon", "coordinates": [ [ [ [ -83.671771983605112, 33.678568396532036 ], [ -83.671771983605112, 33.678408873108779 ], [ -83.674091523007107, 33.678408873108779 ], [ -83.674091523007107, 33.678571387593394 ], [ -83.671771983605112, 33.678568396532036 ] ] ] ] } },
+{ "type": "Feature", "properties": { "id": null, "asset_type": "combiner", "asset_name": "CMB-001-09" }, "geometry": { "type": "MultiPolygon", "coordinates": [ [ [ [ -83.671771983605069, 33.678376968388598 ], [ -83.671771983605069, 33.678230405928218 ], [ -83.674098711662268, 33.678230405928218 ], [ -83.674098711662268, 33.678396908840099 ], [ -83.671771983605069, 33.678376968388598 ] ] ] ] } },
+{ "type": "Feature", "properties": { "id": null, "asset_type": "combiner", "asset_name": "CMB-001-10" }, "geometry": { "type": "MultiPolygon", "coordinates": [ [ [ [ -83.671555125840754, 33.678219438659241 ], [ -83.671555125840754, 33.678047950271342 ], [ -83.67409032489789, 33.678047950271342 ], [ -83.67409032489789, 33.678216821469839 ], [ -83.671555125840754, 33.678219438659241 ] ] ] ] } }
+]
+}
