+++ demos/data/digitized_plant_meters.geojson
@@ -1,0 +1,16 @@
+{
+"type": "FeatureCollection",
+"name": "digitized_plant_meters",
+"features": [
+{ "type": "Feature", "properties": { "id": null, "asset_type": "combiner", "asset_name": "CMB-001-01" }, "geometry": { "type": "MultiPolygon", "coordinates": [ [ [ [ 99.545680128061662, 241.961353002342094 ], [ 99.545863964293886, 224.267895565563975 ], [ 39.996105137144333, 224.488630328934789 ], [ 40.218288078217896, 228.690830235072212 ], [ 60.216329095194929, 229.133271679172282 ], [ 60.216247081768252, 242.182193106656285 ], [ 99.545680128061662, 241.961353002342094 ] ] ] ] } },
+{ "type": "Feature", "properties": { "id": null, "asset_type": "combiner", "asset_name": "CMB-001-02" }, "geometry": { "type": "MultiPolygon", "coordinates": [ [ [ [ 118.432933184538442, 222.719931349530469 ], [ 118.433315937651514, 191.75629798607477 ], [ 0.444402669322856, 192.197904328330935 ], [ 0.666602788327288, 209.670251699193869 ], [ 20.664685807738191, 209.891442712679122 ], [ 20.664658139853401, 222.719221649329427 ], [ 118.432933184538442, 222.719931349530469 ] ] ] ] } },
+{ "type": "Feature", "properties": { "id": null, "asset_type": "combiner", "asset_name": "CMB-001-03" }, "geometry": { "type": "MultiPolygon", "coordinates": [ [ [ [ 138.209273765432698, 189.102532519951524 ], [ 138.209714049278546, 158.581131379932913 ], [ 0.888808453751342, 158.580134571181844 ], [ 0.888805622345575, 189.101535712310096 ], [ 138.209273765432698, 189.102532519951524 ] ] ] ] } },
+{ "type": "Feature", "properties": { "id": null, "asset_type": "combiner", "asset_name": "CMB-001-04" }, "geometry": { "type": "MultiPolygon", "coordinates": [ [ [ [ 157.096929886255026, 156.369722969947162 ], [ 157.097437588834566, 125.405877174854567 ], [ 1.111014412733015, 125.40458932229383 ], [ 1.111010770886167, 156.810775052905512 ], [ 157.096929886255026, 156.369722969947162 ] ] ] ] } },
+{ "type": "Feature", "properties": { "id": null, "asset_type": "combiner", "asset_name": "CMB-001-05" }, "geometry": { "type": "MultiPolygon", "coordinates": [ [ [ [ 176.429124970706255, 123.415677309788336 ], [ 176.429450786479947, 105.72200298949933 ], [ 1.333220031965167, 105.720378690953908 ], [ 1.33321753912263, 123.635223737732744 ], [ 176.429124970706255, 123.415677309788336 ] ] ] ] } },
+{ "type": "Feature", "properties": { "id": null, "asset_type": "combiner", "asset_name": "CMB-001-06" }, "geometry": { "type": "MultiPolygon", "coordinates": [ [ [ [ 176.096196591150289, 102.957357321518217 ], [ 176.096503497951034, 86.258917717529386 ], [ 1.99983411038095, 86.257299665775221 ], [ 1.999830463426523, 103.729838421528029 ], [ 176.096196591150289, 102.957357321518217 ] ] ] ] } },
+{ "type": "Feature", "properties": { "id": null, "asset_type": "combiner", "asset_name": "CMB-001-07" }, "geometry": { "type": "MultiPolygon", "coordinates": [ [ [ [ 196.428209216412682, 83.384082049385214 ], [ 196.428680787722271, 60.38220391250983 ], [ 0.888817562276747, 60.380190439994962 ], [ 0.888815387441941, 83.824411853138344 ], [ 196.428209216412682, 83.384082049385214 ] ] ] ] } },
+{ "type": "Feature", "properties": { "id": null, "asset_type": "combiner", "asset_name": "CMB-001-08" }, "geometry": { "type": "MultiPolygon", "coordinates": [ [ [ [ 215.760522484597374, 57.728553357995708 ], [ 215.760920929485451, 40.034758863547083 ], [ 0.666614587396089, 40.032329544056402 ], [ 0.666613333279225, 58.057882378197434 ], [ 215.760522484597374, 57.728553357995708 ] ] ] ] } },
+{ "type": "Feature", "properties": { "id": null, "asset_type": "combiner", "asset_name": "CMB-001-09" }, "geometry": { "type": "MultiPolygon", "coordinates": [ [ [ [ 215.761000623869592, 36.495996078822756 ], [ 215.761366693818843, 20.239787904754326 ], [ 0.000000002359505, 20.237358561035865 ], [ 0.000000002359501, 38.705293627470105 ], [ 215.761000623869592, 36.495996078822756 ] ] ] ] } },
+{ "type": "Feature", "properties": { "id": null, "asset_type": "combiner", "asset_name": "CMB-001-10" }, "geometry": { "type": "MultiPolygon", "coordinates": [ [ [ [ 235.870978184138409, 19.023809801736316 ], [ 235.871446433731194, 0.002903297413217 ], [ 0.777720270530674, 0.000000036652085 ], [ 0.777718750054046, 18.730616826334078 ], [ 235.870978184138409, 19.023809801736316 ] ] ] ] } }
+]
+}
