+++ docs/sphinx/source/demos/automate_cmv_demo.nblink
@@ -1,0 +1,3 @@
+{
+  "path":"../../../../demos/automate_cmv_demo.ipynb"
+}
