+++ docs/sphinx/source/demos/cmv_demo.nblink
@@ -1,0 +1,3 @@
+{
+  "path":"../../../../demos/cmv_demo.ipynb"
+}
