--- docs/sphinx/source/demos/data/digitized_plant_centroids.csv
@@ -1,11 +1,0 @@
-asset_name,E,N
-CMB-001-01,76.50439339352886,232.4625409910081
-CMB-001-02,63.430048855892466,206.6576786329951
-CMB-001-03,69.54915047353538,173.84132544336165
-CMB-001-04,78.91971734693506,140.9979925797982
-CMB-001-05,88.69999446040173,114.62343072114969
-CMB-001-06,88.39076881801633,94.80230598703163
-CMB-001-07,98.34825037424017,71.99306524359586
-CMB-001-08,107.88070481821165,48.963631280314324
-CMB-001-09,105.59013338994684,28.931320169791586
-CMB-001-10,118.62575746712822,9.439519127662802
