+++ docs/sphinx/source/demos/field_demo_detailed.nblink
@@ -1,0 +1,3 @@
+{
+  "path":"../../../../demos/field_demo_detailed.ipynb"
+}
