+++ docs/sphinx/source/demos/field_reassignment_demo.nblink
@@ -1,0 +1,3 @@
+{
+  "path":"../../../../demos/field_reassignment_demo.ipynb"
+}
