+++ docs/sphinx/source/demos/plant_digitization_example.nblink
@@ -1,0 +1,6 @@
+{
+  "path":"../../../../demos/plant_digitization_example.ipynb",
+  "extra-media":[
+    "../../../../demos/images"
+  ]
+}
